feat: Update nes/gbc/sms/genesis emulators to use shared memory (#99)
* Add `shared_memory` component for sharing memory between the emulators more easily
* Refactored all emulators to use new shared memory for majority of state and data. This means each emulator component now has less than 5k of D/IRAM in use that is not from shared memory.
* Update to have better clear screen function that wont have any race conditions and wont destroy buffers. Update video task to have simple clear screen function if passed nullptr.
* Update cmake lists to use newer version of cmake
* Update to hid roms from rom list if their emulator is not enabled for easier / faster iteration and testing when adding / modifying emulators
* Update espp to latest

Shared memory means that it is not easier to add more emulators (such as msx and doom) since the emulators no longer fill up the internal ram. This does require in some cases modifying the emulators to leverage the shared memory, but this is a one-time operation and actually helps ensure the emulator maintains high performance by keeping its state within internal memory as much as possible, instead of in offboard PSRAM.

Build and run `main` on box-3 and ensure all emulators work for multiple sessions of different roms in sequence. Also ensure that saving state and loading state works for each emulator.